Add select command to use select_prompt and fix issues with it
Useful for testing and allows it to be used in a script.

Also fixed a bug where the wrong result was returned and made it print
the selected result in cyan once one is selected. It doesn't return
if all items are filter and you press enter. It prints things correctly
if there is only one option.