Run create-tauri-app for wkfl
This will be a local app that I build to be paired with the wkfl cli. We
will see what I end up doing with it, but thought it would be
interesting to try and build a mac app.

Also added a new codespaces environment for working on it, since you
need a desktop environment to run it.