Fix issue with select overwriting terminal text of previous command
The select prompt was moving up to far and causing it to overwrite the
previous command. It also made the prompt look weird because it doesn't
rerender some elements of the prompt and it was showing the wrong thing.
I also made it handle being at the bottom of a terminal window and added
a short doc about why we are using stderr.